services/finance-reconciliation/requirements.txt
google-cloud-bigquery==3.25.0
jinja2==3.1.4
